Install ceph-extras package
Also only install if machine is x86_64.

+++ fabfile.py
@@ -91,7 +91,7 @@ def _apt_add_testing_repo(branch):
     sudo('echo deb http://gitbuilder.ceph.com/ceph-deb-$(lsb_release -sc)-x86_64-basic/ref/{branch} $(lsb_release -sc) main | sudo tee /etc/apt/sources.list.d/ceph.list'.format(branch=branch))
 
 def _apt_install(*packages):
-    
+    _ceph_extras()
     sudo("apt-get update")
     sudo(' '.join(
             [
@@ -107,7 +107,8 @@ def _apt_install(*packages):
             + list(packages)))
 
 def _apt_reinstall_for_backports(*packages):
-
+    if 'x86_64' not in env.host_string:
+        return
     sudo("mkdir -p /srv/extras-backports")
     sudo("rm -f /srv/extras-backports/*")
     sudo("apt-get clean")
@@ -920,3 +921,5 @@ def install_git():
                 sudo('make install')
                 sudo('rm -Rf /srv/src/git-{version}'.format(version=git_version))
 
+def _ceph_extras():
+    sudo('echo deb http://ceph.com/packages/ceph-extras/debian $(lsb_release -sc) main | sudo tee /etc/apt/sources.list.d/ceph-extras.list')
